fix key-order bug and move db instantiation to db folder
