+++ app/actions/action_dropNewPin.js
@@ -1,5 +1,5 @@
-import { DROP_NEW_PIN , userData } from '../constants/constants.js';
-
+import { DROP_NEW_PIN } from '../constants/constants.js';
+import { userData } from '../lib/db/db.js';
 function dropNewPin(pinInfo) {
   return {
     type: DROP_NEW_PIN,
@@ -21,14 +21,15 @@ export default function getLocationToSave(location) {
               //this will add id prop w unique id generated by db to each added pin
               userData.on("value", function(snap) {
                 snap.forEach((child) => {
-                  coords.id = key;
                   var key = child.key();
+                  coords.id = key;
                   userData.child(key).update({
                     id: key
                   });
+                  dispatch(dropNewPin(coords));
                 });
               });
-              dispatch(dropNewPin(coords));
+
           },
           (error) => {
             console.error(error);
@@ -43,14 +44,15 @@ export default function getLocationToSave(location) {
               //this will add id prop w unique id generated by db to each added pin
         userData.on("value", function(snap) {
           snap.forEach((child) => {
-            coords.id = key;
             var key = child.key();
+            coords.id = key;
             userData.child(key).update({
               id: key
             });
+            dispatch(dropNewPin(coords));
           });
         });
-      dispatch(dropNewPin(coords));
+
       }
     };
 }
