+++ app/lib/db/db.js
@@ -1,0 +1,3 @@
+const Firebase = require('firebase');
+export const ref = new Firebase("https://interruptedlobster.firebaseio.com/tiff");
+export const userData = ref.child('user')
